requirements.txt
python-telegram-bot>=22.5,<23
aiosqlite>=0.20.0,<1
python-dotenv>=1.0.1,<2
pytest>=8.3.4,<9
pytest-asyncio>=0.24.0,<1
